requirements.txt
opencv-python==4.9.0.80
Pillow==10.3.0
pytesseract==0.3.10
scikit-image==0.23.2
scikit-learn==1.4.2
tensorflow==2.16.1
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.0
seaborn==0.13.2
streamlit==1.35.0
joblib==1.4.2
reportlab==4.2.0
roboflow==1.1.27